Added PauseManager and ScreenManager, changed IObserver and SaveData
Added new scene Mobile PauseManager.
Added new script LocalizationTime, ISreen, PauseManager, PlayerScreen, ScreenManager, PerkSkillsData, DeleteDataButton, GameOverScreen, PauseButton, PauseLevelScreen, PauseMenuScreen, SettingsScreen, WinScreen, ScreenTest, EnemyCounterUI, ShowRewardScreen, LevelSingScreen.

Changed scripts IObserver(changed method Update data for a lot of parameters), all scripts with IObserver, Character(deleted all data with health system), CharacterInput(added input for call Pause), CharacterMovement(deleted waste), MVC_Enemy(deleted a few lines repeated), Destructible(added property for CurrentHealth), SceneTransition(deleted call of courutine from Start), GameOverManager(deleted and changed global), LevelButton(changed call of level), MainMenu(deleted unused lines), EventType(added new types of events), ExperienceSystem(added Load and Save), PerkSkillManager(load and save),PlayerBase(health from PerkSkillManager, reorganaized), SaveData(added more vareables for save), SaveWithJSON(now is songleton), GoldParticle(deleted unused lines), GoldSFX(deleted unused lines), TowerShootSFX(deleted lines), WinLooseSFX(deleted lines), GlobalStamina(reworked), SimpleTower(send current health and max health), TowerHealthBar(reworked), UI(deleted unused methods from IObserver).

+++ Assets/Scripts/Saving/SaveWithJSON.cs
@@ -1,14 +1,17 @@
 using UnityEngine;
 using System.IO;
 
-public class SaveWithJSON : MonoBehaviour
+public class SaveWithJSON : SingltonBase<SaveWithJSON>
 {
-    [SerializeField] protected SaveData _saveData = new SaveData();
+    public SaveData _saveData = new SaveData();
     [SerializeField] protected string _pathBase = "";
     [SerializeField] protected string _path = "";
+    private bool _deleted = false;
 
-    private void Awake()
+    protected override void Awake()
     {
+        base.Awake();
+
 #if UNITY_EDITOR
         _pathBase = Application.dataPath;
 #else
@@ -16,53 +19,50 @@ public class SaveWithJSON : MonoBehaviour
 #endif
 
         _path = _pathBase + "/SaveData.Json"; //name of saveData
-    }
-
-    private void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Escape)) SaveGame();       
-        if (Input.GetKeyDown(KeyCode.Space)) LoadGame();       
+        LoadGame();
     }
 
     public void SaveGame()
     {
         var json = JsonUtility.ToJson(_saveData, true);
         File.WriteAllText(_path, json);
-        //PlayerPrefs.SetString("MyJson", json);
     }
 
     public void LoadGame()
     {
-        if(Directory.Exists(_pathBase) && File.Exists(_path))
+        if (!File.Exists(_path))
+        {
+            _saveData = new SaveData();
+            SaveGame();
+            return;
+        }
+
+        if (Directory.Exists(_pathBase) && File.Exists(_path))
         {
             string json = File.ReadAllText(_path);
             JsonUtility.FromJsonOverwrite(json, _saveData);
         }
-
-        //if(PlayerPrefs.HasKey("MyJson"))
-        //{
-        //    string json = PlayerPrefs.GetString("MyJson");
-        //    JsonUtility.FromJsonOverwrite(json, _saveData);
-        //}
     }
 
     public void DeleteAll()
     {
-        File.Delete(_path);
+        _deleted = true;
+
+        if (File.Exists(_path))
+            File.Delete(_path);
+
+        Debug.Log("Deleted save file.");
     }
 
     private void OnApplicationQuit()
     {
-        SaveGame();
+        if (!_deleted)
+            SaveGame();
     }
 
     private void OnApplicationFocus(bool focus)
     {
-        if(!focus) SaveGame();
-    }
-
-    private void OnApplicationPause(bool pause)
-    {
-        if(pause) SaveGame();
+        if (!focus && !_deleted)
+            SaveGame();
     }
 }
